Fix bail!() to be more similar to error-chain
This changes the `bail!()` macro such that if it's passed one parameter,
it's expected to be of a type that can be converted to the correct
error type with From. Otherwise, if it has multiple parameters,
they're treated as a format string an its parameters.

Unfortunately this means that if its passed a single string, it will
not work as expected because a string can't convert directly to an error
type - it requires the use of `failure::err_msg()` to format the string.
This differs from error-chain as it has From conversion from string types
to an Error type.