Split backtrace and std features up.
This allows depending on std Fail implementations without depending on
the backtrace crate.

Note that with this, both the 'backtrace' and 'std' features must
be enabled in order to access backtrace functionality. There is no
current way to depend on the 'std' feature from the 'backtrace' feature
in cargo, as it is named after a dependency.