Implement flat error mapping
The most natural way to chain errors in the failure crate is to
have failures that contain other Errors. This is supported in the
derive(Fail) macro using fail(cause). However, confusingly, this
doesn't work for failure::Error!

This PR adds a new directive, flat_cause, that acts as a flat_map,
taking the underlying Fail out of a failure::Error.

As an added bonus, it works with any type that implements
cause(&self) -> &Fail.