Add an is_empty method to Backtrace
Fixes #162.

The purpose of this method is to enable testing for when printing a
backtrace will be an empty string because the environment variable
wasn't enabled without needing to Display the backtrace and test for
empty string.

Always returns true in no_std mode because there isn't a backtrace so
the Display will always be empty string.